web/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
billiard==3.6.4.0
bitarray==2.5.1
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cytoolz==0.11.2
dfktools==0.3.1
# eth-abi==2.1.1
# eth-account==0.5.8
# eth-hash==0.3.2
# eth-keyfile==0.5.1
# eth-keys==0.3.4
eth-rlp==0.3.0
# eth-typing==2.3.0
# eth-utils==1.10.0
# ipfshttpclient==0.8.0a2
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
importlib-resources==5.8.0
jsonschema==4.6.0
kombu==5.2.4
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
parsimonious==0.8.1
prompt-toolkit==3.0.29
protobuf==3.20.1
pycryptodome==3.14.1
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
rlp==2.0.1
six==1.16.0
toolz==0.11.2
urllib3==1.26.9
varint==1.0.2
vine==5.0.0
vine==5.0.0
wcwidth==0.2.5
web3==6.4.0
websockets==9.1
yarl==1.7.2
zipp==3.8.0
SQLAlchemy==1.4.22
psycopg2==2.9.1
cryptography==36.0.2
discord-webhook==0.17.0

